Java-based chess client/server system
Client/server system for playing chess online.  ChessServer must be
running, and the user of ChessClient must know ChessServer's IP address.
The user of ChessClient logs in (or creates an account).  Once logged
in, the user can see all other users logged in to the ChessServer.  The
user may challenge any of them to a game of chess.  The game is played
in the client but controlled by the server; the server will reject all
invalid moves.  If one or both users disconnect, the server will
continue running the game, and will resume it when either player logs
back in.